fix(verify): miscellaneous fixes for home-base e2e test

+++ platform_umbrella/apps/verify/lib/verify/kind_install_worker.ex
@@ -5,7 +5,6 @@ defmodule Verify.KindInstallWorker do
 
   alias CommonCore.ApiVersionKind
   alias CommonCore.Ecto.BatteryUUID
-  alias CommonCore.StateSummary.Namespaces
   alias Verify.PathHelper
 
   require Logger
@@ -32,9 +31,13 @@ defmodule Verify.KindInstallWorker do
     state = struct!(State, args)
 
     Logger.info("Starting KindInstallWorker")
+    Process.flag(:trap_exit, true)
 
     {:ok, state}
   end
+
+  def terminate(_reason, %{bi_binary: nil} = _state), do: :ok
+  def terminate(_reason, state), do: do_stop_all(state)
 
   # determine bi_binary if necessary
   def handle_call({:start, args}, _from, %{bi_binary: nil} = state) do
@@ -54,28 +57,52 @@ defmodule Verify.KindInstallWorker do
     do_stop_all(state)
   end
 
-  def handle_call({:rage, {mod, name}}, _from, %{bi_binary: nil} = state) do
-    do_rage(mod, name, %{state | bi_binary: PathHelper.find_bi()})
+  def handle_call({:rage, output}, _from, %{bi_binary: nil} = state) do
+    do_rage(output, %{state | bi_binary: PathHelper.find_bi()})
   end
 
-  def handle_call({:rage, {mod, output}}, _from, state) do
-    do_rage(mod, output, state)
+  def handle_call({:rage, output}, _from, state) do
+    do_rage(output, state)
   end
 
-  defp do_start({mod, identifier, slug}, state) do
-    {spec, path} = build_install_spec(identifier, slug, state)
+  defp do_start({:cmd, cmd, slug}, state) do
+    Logger.debug("Running #{cmd}")
+
+    latest_release =
+      [{Tesla.Middleware.FollowRedirects, max_redirects: 3}]
+      |> Tesla.client()
+      |> Tesla.get!("https://api.github.com/repos/batteries-included/batteries-included/releases/latest")
+      |> Map.get(:body)
+      |> Jason.decode!()
+      |> Map.get("name")
+
+    Logger.debug("Using latest release: #{latest_release}")
+
+    env = [
+      {"BI_VERSION_TAG", latest_release},
+      {"BI_IMAGE_TAR", System.get_env("BI_IMAGE_TAR", "")},
+      {"VERSION_OVERRIDE", System.get_env("VERSION_OVERRIDE", "")}
+    ]
+
+    common_start(fn -> System.shell(cmd, env: env, stderr_to_stdout: true, lines: 1024) end, state, slug)
+  end
+
+  defp do_start({:spec, identifier, slug}, state) do
+    {_spec, path} = build_install_spec(identifier, slug, state)
     Logger.debug("Starting with #{path}")
     env = [{"BI_IMAGE_TAR", System.get_env("BI_IMAGE_TAR", "")}]
 
-    with {_output, 0} <-
-           System.cmd(state.bi_binary, ["start", path], env: env),
-         {kube_config_path, 0} <-
-           System.cmd(state.bi_binary, ["debug", "kube-config-path", path]),
-         {:ok, url} <- get_url(spec, kube_config_path) do
-      Logger.debug("Kind install started from #{path}")
+    common_start(fn -> System.cmd(state.bi_binary, ["start", path], env: env) end, state, slug, path)
+  end
+
+  defp common_start(func, state, slug, path \\ "") do
+    with {_output, 0} <- func.(),
+         {kube_config_path, 0} <- System.cmd(state.bi_binary, ["debug", "kube-config-path", slug]),
+         {:ok, url} <- get_url(kube_config_path) do
+      Logger.debug("Kind install started for #{slug}")
       Logger.debug("Kubeconfig found at #{kube_config_path}")
 
-      {:reply, {:ok, url, kube_config_path}, %{state | started: Map.put(state.started, mod, path)}}
+      {:reply, {:ok, url, kube_config_path}, %{state | started: Map.put(state.started, slug, path)}}
     else
       error ->
         Logger.warning("Unable to start Kind install from #{path}")
@@ -83,39 +110,46 @@ defmodule Verify.KindInstallWorker do
     end
   end
 
-  defp do_stop_all(%{started: started} = state) do
-    Enum.each(started, fn {_, path} ->
-      :ok = do_stop(state.bi_binary, path)
+  defp do_stop_all(%{started: started, bi_binary: bi} = state) do
+    Enum.each(started, fn {slug, path} ->
+      :ok = do_stop(slug, path, bi)
     end)
 
     {:reply, :ok, %{state | started: %{}}}
   end
 
-  defp do_stop(bi, path) do
-    spec = path |> File.read!() |> Jason.decode!()
-    slug = Map.fetch!(spec, "slug")
-
+  defp do_stop(slug, path, bi) do
     Logger.info("Stopping Kind install with path: #{path}, slug: #{slug}, bi: #{bi}")
 
     {_, 0} = System.cmd(bi, ["stop", slug])
 
     # Remove the file after stopping the install
-    _ = File.rm_rf(path)
+    if path != "" do
+      _ = File.rm_rf(path)
+    end
+
     :ok
   end
 
-  defp do_rage(mod, output, %{bi_binary: bi, started: started} = state) do
-    {_, path} = Enum.find(started, fn {module, _} -> module == mod end)
+  defp do_rage(output, %{bi_binary: bi, started: started} = state) do
+    time = :second |> :erlang.system_time() |> to_string()
 
-    case System.cmd(bi, ["rage", path, "-o=#{output}"], stderr_to_stdout: true) do
-      {output, 0} ->
-        Logger.debug("Rage ran: #{inspect(output)}")
-        {:reply, :ok, state}
+    results =
+      Enum.map(started, fn {slug, path} ->
+        full_output = Path.join(output, "#{time}_#{slug}.json")
 
-      response ->
-        Logger.error("Rage failed for: #{inspect(response)}")
-        {:reply, response, state}
-    end
+        case System.cmd(bi, ["rage", path, "-o=#{full_output}"], stderr_to_stdout: true) do
+          {stdout, 0} ->
+            Logger.debug("Rage ran: #{inspect(stdout)}")
+            :ok
+
+          response ->
+            Logger.error("Rage failed for: #{inspect(response)}")
+            response
+        end
+      end)
+
+    {:reply, results, state}
   end
 
   def build_install_spec(identifier, slug, %{root_path: root_dir} = _state) do
@@ -132,13 +166,12 @@ defmodule Verify.KindInstallWorker do
     {spec, path}
   end
 
-  def get_url(%{target_summary: summary} = _spec, kube_config_path) do
+  def get_url(kube_config_path) do
     # don't use the connection pool
     {:ok, conn} = K8s.Conn.from_file(kube_config_path, insecure_skip_tls_verify: true)
 
     {api_version, kind} = ApiVersionKind.from_resource_type!(:config_map)
-    core_namespace = Namespaces.core_namespace(summary)
-    op = K8s.Client.get(api_version, kind, name: "access-info", namespace: core_namespace)
+    op = K8s.Client.get(api_version, kind, name: "access-info", namespace: "battery-core")
 
     case K8s.Client.run(conn, op) do
       {:ok, %{"data" => %{"hostname" => hostname, "ssl" => ssl}}} ->
@@ -149,15 +182,19 @@ defmodule Verify.KindInstallWorker do
     end
   end
 
-  def start(mod, identifier, slug) do
-    GenServer.call(__MODULE__, {:start, {mod, identifier, slug}}, 15 * 60 * 1000)
+  def start_from_command(target, start_cmd, slug) do
+    GenServer.call(target, {:start, {:cmd, start_cmd, slug}}, 15 * 60 * 1000)
   end
 
-  def rage(mod, output) do
-    GenServer.call(__MODULE__, {:rage, {mod, output}}, 15 * 60 * 1000)
+  def start_from_spec(target, identifier, slug) do
+    GenServer.call(target, {:start, {:spec, identifier, slug}}, 15 * 60 * 1000)
   end
 
-  def stop_all do
-    GenServer.call(__MODULE__, :stop_all, 15 * 60 * 1000)
+  def rage(target, output) do
+    GenServer.call(target, {:rage, output}, 15 * 60 * 1000)
+  end
+
+  def stop_all(target) do
+    GenServer.call(target, :stop_all, 15 * 60 * 1000)
   end
 end
